custom-element-registry/index.html

<!doctype html>

<meta charset=utf-8>

<title>Custom Elements: CustomElementRegistry interface</title>

<meta name=author title=snuggs href=//github.com/snuggs>
<meta name=assert content='CustomElementRegistry interface must exist'>

<link rel=help href=https://www.w3.org/TR/custom-elements/>
<link rel=help href=https://html.spec.whatwg.org/multipage/custom-elements.html>
<link rel=help href=https://github.com/w3c/web-platform-tests/blob/master/custom-elements/CustomElementRegistry.html>

<script defer src=/resources/testharness.js></script>
<script defer src=/resources/testharnessreport.js></script>

<!-- Testing Format https://wiki.csswg.org/test/format#reference-links -->

<script>

test(function () {
    assert_true('define' in CustomElementRegistry.prototype, '"define" exists on CustomElementRegistry.prototype');
    assert_true('define' in customElements, '"define" exists on window.customElements');
}, 'CustomElementRegistry interface must have define as a method');

test(function () {
    assert_throws({'name': 'TypeError'}, function () { customElements.define('badname', 1); },
        'customElements.define must throw a TypeError when the element interface is a number');
    assert_throws({'name': 'TypeError'}, function () { customElements.define('badname', '123'); },
        'customElements.define must throw a TypeError when the element interface is a string');
    assert_throws({'name': 'TypeError'}, function () { customElements.define('badname', {}); },
        'customElements.define must throw a TypeError when the element interface is an object');
    assert_throws({'name': 'TypeError'}, function () { customElements.define('badname', []); },
        'customElements.define must throw a TypeError when the element interface is an array');
}, 'customElements.define must throw when the element interface is not a constructor');

test(function () {
    customElements.define('custom-html-element', HTMLElement);
}, 'customElements.define must not throw the constructor is HTMLElement');

</script>
